@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Collections.Specialized;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -12,47 +11,46 @@ using Http.Resilience.Extensions;
 namespace Http.Resilience.Internals
 {
     /// <summary>
-    /// Defines the options used for configuring the retry policy.
+    ///     Defines the options used for configuring the retry policy.
     /// </summary>
     public class HttpRetryOptions
     {
         /// <summary>
-        /// Returns false if we should continue retrying based on the response, and true
-        /// if we should not, even though this is technically a retryable status code.
+        ///     Returns false if we should continue retrying based on the response, and true
+        ///     if we should not, even though this is technically a retryable status code.
         /// </summary>
         /// <param name="response">The response to check if we should retry the request.</param>
         /// <returns>False if we should retry, true if we should not based on the response.</returns>
-        public delegate bool HttpResponseMessageFilter(int statusCode, IEnumerable<KeyValuePair<string, IEnumerable<string>>> headers);
-
-        private int isReadOnly;
-        private int maxRetries;
-        private TimeSpan minBackoff;
-        private TimeSpan maxBackoff;
-        private TimeSpan backoffCoefficient;
-        private ICollection<HttpStatusCode> retryableStatusCodes;
-        private ICollection<HttpResponseMessageFilter> httpResponseMessageFilters;
-        private bool ensureSuccessStatusCode = true;
+        public delegate bool HttpResponseMessageFilter(int statusCode,
+            IEnumerable<KeyValuePair<string, IEnumerable<string>>> headers);
 
         private static readonly TimeSpan DefaultMinBackoff = TimeSpan.FromSeconds(1d);
         private static readonly TimeSpan DefaultMaxBackoff = TimeSpan.FromSeconds(10d);
         private static readonly TimeSpan DefaultBackoffCoefficient = TimeSpan.FromSeconds(1d);
         private static readonly int DefaultMaxRetries = 5;
 
-        private static readonly Lazy<HttpRetryOptions> DefaultOptions = new Lazy<HttpRetryOptions>(() =>
+        private static readonly Lazy<HttpRetryOptions> DefaultOptions = new(() =>
         {
             return new HttpRetryOptions().MakeReadonly();
         });
 
-        private static readonly HttpResponseMessageFilter hostShutdownFilter = (statusCode, headers) =>
+        private static readonly HttpResponseMessageFilter HostShutdownFilter = (statusCode, headers) =>
         {
             return headers.Any(h => h.Key == "X-VSS-HostOfflineError");
         };
 
+        private TimeSpan backoffCoefficient;
+        private bool ensureSuccessStatusCode = true;
+        private ICollection<HttpResponseMessageFilter> httpResponseMessageFilters;
+
+        private int isReadOnly;
+        private TimeSpan maxBackoff;
+        private int maxRetries;
+        private TimeSpan minBackoff;
+        private ICollection<HttpStatusCode> retryableStatusCodes;
+
         public HttpRetryOptions()
-            : this(new HttpResponseMessageFilter[1]
-            {
-                hostShutdownFilter
-            })
+            : this(new HttpResponseMessageFilter[1] { HostShutdownFilter })
         {
         }
 
@@ -64,21 +62,19 @@ namespace Http.Resilience.Internals
             this.MaxRetries = DefaultMaxRetries;
             this.RetryableStatusCodes = new HashSet<HttpStatusCode>
             {
-                HttpStatusCode.BadGateway,
-                HttpStatusCode.GatewayTimeout,
-                HttpStatusCode.ServiceUnavailable
+                HttpStatusCode.BadGateway, HttpStatusCode.GatewayTimeout, HttpStatusCode.ServiceUnavailable
             };
             this.httpResponseMessageFilters = new HashSet<HttpResponseMessageFilter>(filters);
         }
 
         /// <summary>
-        /// Gets a singleton read-only instance of the default settings.
+        ///     Gets a singleton read-only instance of the default settings.
         /// </summary>
         public static HttpRetryOptions Default => DefaultOptions.Value;
 
         /// <summary>
-        /// Calls <seealso cref="HttpResponseMessage.EnsureSuccessStatusCode"/> on <seealso cref="HttpResponseMessage"/>
-        /// which converts an unsuccessful HTTP status code into a <seealso cref="HttpRequestException"/> (Default=true).
+        ///     Calls <seealso cref="HttpResponseMessage.EnsureSuccessStatusCode" /> on <seealso cref="HttpResponseMessage" />
+        ///     which converts an unsuccessful HTTP status code into a <seealso cref="HttpRequestException" /> (Default=true).
         /// </summary>
         public bool EnsureSuccessStatusCode
         {
@@ -91,7 +87,7 @@ namespace Http.Resilience.Internals
         }
 
         /// <summary>
-        /// Gets or sets the coefficient which exponentially increases the backoff starting at MinBackoff.
+        ///     Gets or sets the coefficient which exponentially increases the backoff starting at MinBackoff.
         /// </summary>
         public TimeSpan BackoffCoefficient
         {
@@ -104,7 +100,7 @@ namespace Http.Resilience.Internals
         }
 
         /// <summary>
-        /// Gets or sets the minimum backoff interval to be used.
+        ///     Gets or sets the minimum backoff interval to be used.
         /// </summary>
         public TimeSpan MinBackoff
         {
@@ -117,7 +113,7 @@ namespace Http.Resilience.Internals
         }
 
         /// <summary>
-        /// Gets or sets the maximum backoff interval to be used.
+        ///     Gets or sets the maximum backoff interval to be used.
         /// </summary>
         public TimeSpan MaxBackoff
         {
@@ -129,17 +125,11 @@ namespace Http.Resilience.Internals
             }
         }
 
-        // TODO:  Note that this was renamed from
-        //  "maxRetries" to match the behavior of the parameter (i.e. maxRetries was previously
-        //  behaving like maxAttempts).
-
-
         /// <summary>
-        /// Gets or sets the maximum number of retries allowed.
+        ///     Gets or sets the maximum number of retries for one particular request.
         /// </summary>
         /// <remarks>
-        /// This is the total number of attempts to invoke the submitted action with. A value of
-        //  1 indicates that no retries will be attempted.
+        ///     A value of 0 indicates that no retries will be attempted.
         /// </remarks>
         public int MaxRetries
         {
@@ -152,7 +142,7 @@ namespace Http.Resilience.Internals
         }
 
         /// <summary>
-        /// Gets a set of HTTP status codes which should be retried.
+        ///     Gets a set of HTTP status codes which should be retried.
         /// </summary>
         public ICollection<HttpStatusCode> RetryableStatusCodes
         {
@@ -165,48 +155,69 @@ namespace Http.Resilience.Internals
         }
 
         /// <summary>
-        /// How to verify that the response can be retried.
+        ///     Verifies if the request with the given <paramref name="httpResponseMessage" /> can be retried.
         /// </summary>
         /// <param name="httpResponseMessage">Response message from a request.</param>
         /// <returns>True if the request can be retried, false otherwise.</returns>
         public bool IsRetryableResponse(HttpResponseMessage httpResponseMessage)
         {
-            if (this.retryableStatusCodes.Contains(httpResponseMessage.StatusCode))
+            var hasRetryableStatusCode = this.retryableStatusCodes.Contains(httpResponseMessage.StatusCode);
+
+            var isRetryFiltered = this.IsRetryFiltered(httpResponseMessage);
+
+            return hasRetryableStatusCode && !isRetryFiltered;
+        }
+
+
+        public bool IsRetryableResponse(HttpWebResponse httpWebResponse)
+        {
+            var hasRetryableStatusCode = this.retryableStatusCodes.Contains(httpWebResponse.StatusCode);
+
+            var isRetryFiltered = this.IsRetryFiltered(httpWebResponse);
+
+            return hasRetryableStatusCode && !isRetryFiltered;
+        }
+
+        private bool IsRetryFiltered(HttpResponseMessage httpResponseMessage)
+        {
+            foreach (var filter in this.httpResponseMessageFilters)
             {
-                foreach (var filter in this.httpResponseMessageFilters)
+                try
                 {
                     if (filter((int)httpResponseMessage.StatusCode, httpResponseMessage.Headers))
                     {
-                        return false;
+                        return true;
                     }
                 }
-
-                return true;
+                catch
+                {
+                }
             }
 
             return false;
         }
 
-        public bool IsRetryableResponse(HttpWebResponse httpWebResponse)
+        private bool IsRetryFiltered(HttpWebResponse httpWebResponse)
         {
-            if (this.retryableStatusCodes.Contains(httpWebResponse.StatusCode))
+            foreach (var filter in this.httpResponseMessageFilters)
             {
-                foreach (var filter in this.httpResponseMessageFilters)
+                try
                 {
                     if (filter((int)httpWebResponse.StatusCode, httpWebResponse.Headers.GetHeaders()))
                     {
-                        return false;
+                        return true;
                     }
                 }
-
-                return true;
+                catch
+                {
+                }
             }
 
             return false;
         }
 
         /// <summary>
-        /// Ensures that no further modifications may be made to the retry options.
+        ///     Ensures that no further modifications may be made to the retry options.
         /// </summary>
         /// <returns>A read-only instance of the retry options</returns>
         public HttpRetryOptions MakeReadonly()
@@ -214,20 +225,22 @@ namespace Http.Resilience.Internals
             if (Interlocked.CompareExchange(ref this.isReadOnly, 1, 0) == 0)
             {
                 this.retryableStatusCodes = new ReadOnlyCollection<HttpStatusCode>(this.retryableStatusCodes.ToList());
-                this.httpResponseMessageFilters = new ReadOnlyCollection<HttpResponseMessageFilter>(this.httpResponseMessageFilters.ToList());
+                this.httpResponseMessageFilters =
+                    new ReadOnlyCollection<HttpResponseMessageFilter>(this.httpResponseMessageFilters.ToList());
             }
 
             return this;
         }
 
         /// <summary>
-        /// Throws an InvalidOperationException if this is marked as ReadOnly.
+        ///     Throws an InvalidOperationException if this is marked as ReadOnly.
         /// </summary>
         private void ThrowIfReadonly([CallerMemberName] string propertyName = "")
         {
             if (this.isReadOnly > 0)
             {
-                throw new InvalidOperationException($"{nameof(HttpRetryOptions)} is marked as readonly; '{propertyName}' cannot be changed.");
+                throw new InvalidOperationException(
+                    $"{nameof(HttpRetryOptions)} is marked as readonly; '{propertyName}' cannot be changed.");
             }
         }
     }
