@@ -2,25 +2,22 @@ using System;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Http.Resilience.Internals;
-using Http.Resilience.Internals.Logging;
+using Http.Resilience.Logging;
 
 namespace Http.Resilience
 {
     /// <summary>
-    /// HTTP retry helper which can be used to retry failed HTTP calls.
-    /// The retry behavior is configurable in <seealso cref="HttpRetryOptions"/>.
+    ///     HTTP retry helper which can be used to retry failed HTTP calls.
+    ///     The retry behavior is configurable in <seealso cref="HttpRetryOptions" />.
     /// </summary>
-    public class HttpRetryHelper
+    public class HttpRetryHelper : IHttpRetryHelper
     {
         private readonly string instance = Guid.NewGuid().ToString().Substring(0, 5).ToUpperInvariant();
-        private readonly HttpRetryOptions options;
 
         private Func<Exception, bool> canRetryDelegate;
 
-        public HttpRetryOptions Options => this.options;
-
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with default <seealso cref="HttpRetryOptions"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with default <seealso cref="HttpRetryOptions" />.
         /// </summary>
         public HttpRetryHelper()
             : this(HttpRetryOptions.Default)
@@ -28,8 +25,8 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with default <seealso cref="HttpRetryOptions"/>
-        /// overriding <paramref name="maxRetries"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with default <seealso cref="HttpRetryOptions" />
+        ///     overriding <paramref name="maxRetries" />.
         /// </summary>
         public HttpRetryHelper(int maxRetries)
             : this(new HttpRetryOptions { MaxRetries = maxRetries })
@@ -37,15 +34,17 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Creates an instance of <seealso cref="HttpRetryHelper"/> with <paramref name="httpRetryOptions"/>.
+        ///     Creates an instance of <seealso cref="HttpRetryHelper" /> with <paramref name="httpRetryOptions" />.
         /// </summary>
         public HttpRetryHelper(HttpRetryOptions httpRetryOptions)
         {
-            this.options = httpRetryOptions;
+            this.Options = httpRetryOptions;
         }
 
+        public HttpRetryOptions Options { get; }
+
         /// <summary>
-        /// Calls <paramref name="action"/> synchronously.
+        ///     Calls <paramref name="action" /> synchronously.
         /// </summary>
         public void Invoke(Action action)
         {
@@ -57,7 +56,7 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> synchronously and returns <typeparamref name="TResult"/>.
+        ///     Calls <paramref name="function" /> synchronously and returns <typeparamref name="TResult" />.
         /// </summary>
         public TResult Invoke<TResult>(Func<TResult> function)
         {
@@ -68,7 +67,7 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> asynchronously.
+        ///     Calls <paramref name="function" /> asynchronously.
         /// </summary>
         public async Task InvokeAsync(Func<Task> function)
         {
@@ -80,13 +79,12 @@ namespace Http.Resilience
         }
 
         /// <summary>
-        /// Calls <paramref name="function"/> asynchronously and returns <typeparamref name="TResult"/>.
+        ///     Calls <paramref name="function" /> asynchronously and returns <typeparamref name="TResult" />.
         /// </summary>
         public async Task<TResult> InvokeAsync<TResult>(Func<Task<TResult>> function)
         {
             var currentAttempt = 1;
-            var maxAttempts = this.options.MaxRetries + 1;
-            var lastStatusCode = 0;
+            var maxAttempts = this.Options.MaxRetries + 1;
             var lastResult = default(TResult);
 
             while (true)
@@ -100,35 +98,22 @@ namespace Http.Resilience
                     lastResult = await function();
                     if (lastResult is HttpResponseMessage httpResponseMessage)
                     {
-                        lastStatusCode = (int)httpResponseMessage.StatusCode;
-
-                        if (this.options.EnsureSuccessStatusCode)
-                        {
-                            httpResponseMessage.EnsureSuccessStatusCode();
-                        }
-                        else
-                        {
-                            if (HasRemainingAttempts(remainingAttempts) && 
-                                this.options.IsRetryableResponse(httpResponseMessage))
-                            {
-                                await this.SleepAsync(remainingAttempts);
-                                currentAttempt++;
-                                this.Log(LogLevel.Info, $"InvokeAsync --> Retry on HttpResponseMessage.StatusCode={(int)httpResponseMessage.StatusCode} ({httpResponseMessage.StatusCode})");
-                                continue;
-                            }
-                        }
+                        httpResponseMessage.EnsureSuccessStatusCode();
                     }
 
-                    this.Log(currentAttempt <= 1 ? LogLevel.Debug : LogLevel.Info, $"InvokeAsync finished successfully (Attempt {currentAttempt}/{maxAttempts})");
+                    this.Log(currentAttempt <= 1 ? LogLevel.Debug : LogLevel.Info,
+                        $"InvokeAsync finished successfully (Attempt {currentAttempt}/{maxAttempts})");
                     return lastResult;
                 }
                 catch (Exception ex)
                 {
-                    this.Log(LogLevel.Error, $"InvokeAsync failed with exception (Attempt {currentAttempt}/{maxAttempts})");
+                    this.Log(LogLevel.Error,
+                        $"InvokeAsync failed with exception (Attempt {currentAttempt}/{maxAttempts})");
 
                     var lastHttpResponseMessage = lastResult as HttpResponseMessage;
-                    if (HasRemainingAttempts(remainingAttempts) && 
-                        (NetworkHelper.IsTransientNetworkException(ex, lastHttpResponseMessage, this.options) || this.canRetryDelegate != null && this.canRetryDelegate(ex)))
+                    if (HasRemainingAttempts(remainingAttempts) &&
+                        (NetworkHelper.IsTransientNetworkException(ex, lastHttpResponseMessage, this.Options) ||
+                         (this.canRetryDelegate != null && this.canRetryDelegate(ex))))
                     {
                         await this.SleepAsync(remainingAttempts);
                         currentAttempt++;
@@ -136,12 +121,51 @@ namespace Http.Resilience
                         continue;
                     }
 
+                    if (lastResult is HttpResponseMessage httpResponseMessage && !this.Options.EnsureSuccessStatusCode)
+                    {
+                        return lastResult;
+                    }
+
                     throw;
                 }
             }
         }
 
-        private static bool HasRemainingAttempts(int remainingAttempts) => remainingAttempts > 0;
+        /// <summary>
+        ///     Custom retry decision logic if an exception occurred.
+        /// </summary>
+        public HttpRetryHelper RetryOnException(Func<Exception, bool> handler)
+        {
+            if (this.canRetryDelegate != null)
+            {
+                throw new InvalidOperationException($"{nameof(RetryOnException)} cannot be called more than once");
+            }
+
+            this.canRetryDelegate = handler;
+            return this;
+        }
+
+        /// <summary>
+        ///     Custom retry decision logic if an exception of type <typeparamref name="TException" /> occurred.
+        /// </summary>
+        public HttpRetryHelper RetryOnException<TException>(Func<TException, bool> handler) where TException : Exception
+        {
+            return this.RetryOnException(ex =>
+            {
+                if (ex is TException tex)
+                {
+                    return handler(tex);
+                }
+
+                return false;
+            });
+        }
+
+        private static bool HasRemainingAttempts(int remainingAttempts)
+        {
+            return remainingAttempts > 0;
+        }
+
         private static int CalculateRemainingAttempts(int currentAttempt, int maxAttempts)
         {
             return maxAttempts - currentAttempt;
@@ -159,37 +183,8 @@ namespace Http.Resilience
 
         protected virtual TimeSpan CalculateBackoff(int remainingAttempts)
         {
-            return BackoffTimerHelper.GetExponentialBackoff(this.options.MaxRetries - remainingAttempts + 1, this.options.MinBackoff, this.options.MaxBackoff, this.options.BackoffCoefficient);
-        }
-
-        /// <summary>
-        /// Custom retry decision logic if an exception occurred.
-        /// </summary>
-        public HttpRetryHelper RetryOnException(Func<Exception, bool> handler)
-        {
-            if (this.canRetryDelegate != null)
-            {
-                throw new InvalidOperationException($"{nameof(RetryOnException)} cannot be called more than once");
-            }
-
-            this.canRetryDelegate = handler;
-            return this;
-        }
-
-        /// <summary>
-        /// Custom retry decision logic if an exception of type <typeparamref name="TException"/> occurred.
-        /// </summary>
-        public HttpRetryHelper RetryOnException<TException>(Func<TException, bool> handler) where TException : Exception
-        {
-            return this.RetryOnException((ex) =>
-            {
-                if (ex is TException tex)
-                {
-                    return handler(tex);
-                }
-
-                return false;
-            });
+            return BackoffTimerHelper.GetExponentialBackoff(this.Options.MaxRetries - remainingAttempts + 1,
+                this.Options.MinBackoff, this.Options.MaxBackoff, this.Options.BackoffCoefficient);
         }
 
         private void Log(LogLevel logLevel, string message)
